Add the change login modal and feature.
Add a method in LoginService to refresh the connected user from the DB

Fix UserService#changeUserPassword not doing the change on a copied user object. (the changeUserLogin method is added with the same behavior). The app will have to call the LoginService#refreshConnectedUser to fetch an updated user

Fix possible race conditions in LoginService (synchronize on "this" when accessing and modifying the loggedUser ref)